+++ lib/components/base-control/container.jsx
@@ -44,7 +44,7 @@ export const Container = forwardRef((props, ref) => {
 				false: 'es:rounded-sm',
 			},
 			compact: {
-				false: 'es:py-2.5 es:min-h-13.5',
+				false: 'es:py-2 es:min-h-13',
 				true: 'es:py-1 es:min-h-9',
 			},
 		},